requirements.txt
aiogram
aiohttp
aiosignal
alembic
async-timeout
attrs
Babel
certifi
charset-normalizer
frozenlist
greenlet
idna
loguru
Mako
MarkupSafe
multidict
python-decouple
pytz
SQLAlchemy
yarl
